add page for emul test
