+++ apps/app/src/app/(protected)/events/create/page.tsx
@@ -1,10 +1,6 @@
-import Link from 'next/link';
-import React from 'react';
 import { CreateEventForm } from './form';
 
-type Props = {};
-
-export default function CreateEvent({}: Props) {
+export default function CreateEvent() {
   return (
     <div className='flex flex-col gap-4 w-full'>
       <h1 className='text-3xl'>Create a new event</h1>
